# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is server Assignment ...",
        "version": "1.0.0",
        "title": "Assignment",
        "host": "localhost:8080",
        "basePath": "/",
        "schemes": [
            "http"
        ]
    },
    "paths": {
...
        "/customer/create-booking": {
            "post": {
                "tags": [
                    "customers"
                ],
                "summary": "Create Booking.",
                "description": "",
                "operationId": "createBooking",
                "consumes": "application/json",
                "produces": "application/json",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "token",
                        "description": "Enter customer token.",
                        "required": true
                    },
                    {
                        "in": "formData",
                        "name": "source",
                        "description": "Enter pick-up point.",
                        "required": true
                    },
                    {
                        "in": "formData",
                        "name": "destination",
                        "description": "Enter destination.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                },
                "security": {}
            }
        },
        "/customer/view-bookings": {
            "post": {
                "tags": [
                    "customers"
                ],
                "summary": "View completed bookings.",
                "description": "",
                "operationId": "viewBooking",
                "consumes": "application/json",
                "produces": "application/json",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "token",
                        "description": "Enter customer token.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                },
                "security": {}
            }
        },
...
        "/admin/view-free-drivers": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "View free drivers.",
                "description": "",
                "operationId": "viewDrivers",
                "consumes": "application/json",
                "produces": "application/json",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "email",
                        "description": "Enter email.",
                        "required": true
                    },
                    {
                        "in": "formData",
                        "name": "password",
                        "description": "Enter password.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                },
                "security": {}
            }
        },
...
        "/driver/login": {
            "post": {
                "tags": [
                    "driver"
                ],
                "summary": "Login to drivers.",
                "description": "",
                "operationId": "driverLogin",
                "consumes": "application/json",
                "produces": "application/json",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "email",
                        "description": "Enter email.",
                        "required": true
                    },
                    {
                        "in": "formData",
                        "name": "password",
                        "description": "Enter password.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                },
                "security": {}
            }
        },
...
    }
}